chore: email test refactor to use fixtures
